+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `hideGameMode` config option
+
 ## [1.5.20] - 2025-07-02
 ### Added
 - Prevent allays from interacting with vanished players